requirements.txt
langdetect==1.0.9
matplotlib==3.4.3
numpy==1.21.1
opencv_python==4.5.3.56
Pillow==9.1.0
protobuf==3.20.1
requests==2.25.1
six==1.16.0
tqdm==4.61.2
